@@ -7,7 +7,7 @@
     <groupId>org.projectrainbow</groupId>
     <artifactId>rainbow-parent</artifactId>
     <packaging>pom</packaging>
-    <version>1.11.2-SNAPSHOT</version>
+    <version>1.12-SNAPSHOT</version>
     <modules>
         <module>install-dependencies</module>
         <module>rainbow</module>
@@ -18,10 +18,10 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.target>1.6</maven.compiler.target>
-        <maven.compiler.source>1.6</maven.compiler.source>
+        <maven.compiler.target>1.8</maven.compiler.target>
+        <maven.compiler.source>1.8</maven.compiler.source>
         <build.number>unknown</build.number>
-        <minecraft.version>1.11.2</minecraft.version>
+        <minecraft.version>1.12</minecraft.version>
     </properties>
 
     <profiles>
