@@ -1,15 +1,12 @@
 package org.projectrainbow.commands;
 
+import PluginReference.MC_Command;
 import PluginReference.MC_Player;
 import PluginReference.RainbowUtils;
 import com.google.common.io.Files;
-import net.minecraft.command.CommandBase;
-import net.minecraft.command.CommandException;
-import net.minecraft.command.ICommandManager;
-import net.minecraft.command.ICommandSender;
-import net.minecraft.server.MinecraftServer;
-import org.projectrainbow._ColorHelper;
 import joebkt._CronData;
+import net.minecraft.command.ICommandManager;
+import org.projectrainbow._ColorHelper;
 import org.projectrainbow._DiwUtils;
 
 import java.io.BufferedInputStream;
@@ -26,12 +23,12 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
-public class _CmdCron extends CommandBase {
+public class _CmdCron implements MC_Command {
     public static Map<String, _CronData> mapData;
     private static String Filename;
 
     static {
-        _CmdCron.mapData = new ConcurrentHashMap<String, _CronData>();
+        _CmdCron.mapData = new ConcurrentHashMap<>();
         _CmdCron.Filename = "Cron.dat";
     }
 
@@ -41,12 +38,22 @@ public class _CmdCron extends CommandBase {
     }
 
     @Override
-    public boolean checkPermission(MinecraftServer minecraftServer, ICommandSender cs) {
-        return (!(cs instanceof MC_Player)) || ((MC_Player) cs).hasPermission("rainbow.cron");
+    public List<String> getAliases() {
+        return Collections.emptyList();
     }
 
     @Override
-    public String getCommandUsage(ICommandSender sender) {
+    public boolean hasPermissionToUse(MC_Player player) {
+        return player == null || player.hasPermission("rainbow.cron");
+    }
+
+    @Override
+    public List<String> getTabCompletionList(MC_Player plr, String[] args) {
+        return Collections.emptyList();
+    }
+
+    @Override
+    public String getHelpLine(MC_Player player) {
         return String.valueOf(_ColorHelper.LIGHT_PURPLE) + "/cron" + _ColorHelper.WHITE + " --- Schedule Commands";
     }
 
@@ -84,7 +91,7 @@ public class _CmdCron extends CommandBase {
             s.close();
         } catch (Throwable exc) {
             System.out.println("Starting New Cron DB: " + _CmdCron.Filename);
-            _CmdCron.mapData = new ConcurrentHashMap<String, _CronData>();
+            _CmdCron.mapData = new ConcurrentHashMap<>();
         }
     }
 
@@ -106,12 +113,12 @@ public class _CmdCron extends CommandBase {
         }
     }
 
-    public void ShowUsage(final ICommandSender cs) {
-        _DiwUtils.reply(cs, String.valueOf(_ColorHelper.RED) + "Usage: /cron " + _ColorHelper.AQUA + "add " + _ColorHelper.YELLOW + "MyLoop " + _ColorHelper.WHITE + "60s " + _ColorHelper.GOLD + "/give @a cake");
-        _DiwUtils.reply(cs, String.valueOf(_ColorHelper.RED) + "Usage: /cron " + _ColorHelper.AQUA + "delete " + _ColorHelper.YELLOW + "MyLoop");
-        _DiwUtils.reply(cs, String.valueOf(_ColorHelper.RED) + "Usage: /cron " + _ColorHelper.AQUA + "setdelay " + _ColorHelper.YELLOW + "MyLoop " + _ColorHelper.WHITE + "5m");
-        _DiwUtils.reply(cs, String.valueOf(_ColorHelper.RED) + "Usage: /cron " + _ColorHelper.AQUA + "setcmd " + _ColorHelper.YELLOW + "MyLoop " + _ColorHelper.GOLD + "/give @a cookie");
-        _DiwUtils.reply(cs, String.valueOf(_ColorHelper.RED) + "Usage: /cron " + _ColorHelper.AQUA + "list");
+    public void ShowUsage(MC_Player player) {
+        _DiwUtils.reply(player, String.valueOf(_ColorHelper.RED) + "Usage: /cron " + _ColorHelper.AQUA + "add " + _ColorHelper.YELLOW + "MyLoop " + _ColorHelper.WHITE + "60s " + _ColorHelper.GOLD + "/give @a cake");
+        _DiwUtils.reply(player, String.valueOf(_ColorHelper.RED) + "Usage: /cron " + _ColorHelper.AQUA + "delete " + _ColorHelper.YELLOW + "MyLoop");
+        _DiwUtils.reply(player, String.valueOf(_ColorHelper.RED) + "Usage: /cron " + _ColorHelper.AQUA + "setdelay " + _ColorHelper.YELLOW + "MyLoop " + _ColorHelper.WHITE + "5m");
+        _DiwUtils.reply(player, String.valueOf(_ColorHelper.RED) + "Usage: /cron " + _ColorHelper.AQUA + "setcmd " + _ColorHelper.YELLOW + "MyLoop " + _ColorHelper.GOLD + "/give @a cookie");
+        _DiwUtils.reply(player, String.valueOf(_ColorHelper.RED) + "Usage: /cron " + _ColorHelper.AQUA + "list");
     }
 
     static long GetMSFromString(String strTime) {
@@ -143,10 +150,10 @@ public class _CmdCron extends CommandBase {
     }
 
     @Override
-    public void execute(MinecraftServer minecraftServer, ICommandSender cs, String[] args) throws CommandException {
+    public void handleCommand(MC_Player player, String[] args) {
         if (args.length >= 1 && args[0].equalsIgnoreCase("add")) {
             if (args.length < 4) {
-                _DiwUtils.reply(cs, _ColorHelper.RED + "Usage: /cron add " + _ColorHelper.YELLOW + "MyLoop " + _ColorHelper.WHITE + "60s " + _ColorHelper.GOLD + "/give @a cake");
+                _DiwUtils.reply(player, _ColorHelper.RED + "Usage: /cron add " + _ColorHelper.YELLOW + "MyLoop " + _ColorHelper.WHITE + "60s " + _ColorHelper.GOLD + "/give @a cake");
             } else {
                 String jobName = args[1].trim();
                 String strTime = args[2].trim();
@@ -154,7 +161,7 @@ public class _CmdCron extends CommandBase {
                 long ms = GetMSFromString(strTime);
                 if (ms > 0L && strCmd.length() > 0) {
                     String key = jobName.toLowerCase();
-                    _CronData data = (_CronData) mapData.get(key);
+                    _CronData data = mapData.get(key);
                     if (data == null) {
                         data = new _CronData();
                     }
@@ -165,43 +172,37 @@ public class _CmdCron extends CommandBase {
                     data.msLastRun = System.currentTimeMillis();
                     mapData.put(key, data);
                     SaveData();
-                    _DiwUtils.reply(cs, _ColorHelper.GREEN + "Saved " + _ColorHelper.AQUA + jobName);
+                    _DiwUtils.reply(player, _ColorHelper.GREEN + "Saved " + _ColorHelper.AQUA + jobName);
                 } else {
-                    _DiwUtils.reply(cs, _ColorHelper.RED + "Usage: /cron add " + _ColorHelper.YELLOW + "MyLoop " + _ColorHelper.WHITE + "60s " + _ColorHelper.GOLD + "/give @a cake");
+                    _DiwUtils.reply(player, _ColorHelper.RED + "Usage: /cron add " + _ColorHelper.YELLOW + "MyLoop " + _ColorHelper.WHITE + "60s " + _ColorHelper.GOLD + "/give @a cake");
                 }
             }
         } else if (args.length >= 1 && args[0].equalsIgnoreCase("delete")) {
             if (args.length <= 1) {
-                _DiwUtils.reply(cs, _ColorHelper.RED + "Usage: /cron " + _ColorHelper.AQUA + "delete " + _ColorHelper.YELLOW + "MyLoop");
+                _DiwUtils.reply(player, _ColorHelper.RED + "Usage: /cron " + _ColorHelper.AQUA + "delete " + _ColorHelper.YELLOW + "MyLoop");
             } else {
                 String jobName = args[1].trim();
                 String key = jobName.toLowerCase();
-                _CronData data = (_CronData) mapData.get(key);
+                _CronData data = mapData.get(key);
                 if (data == null) {
-                    _DiwUtils.reply(cs, _ColorHelper.RED + "No job named: " + _ColorHelper.YELLOW + jobName);
+                    _DiwUtils.reply(player, _ColorHelper.RED + "No job named: " + _ColorHelper.YELLOW + jobName);
                 } else {
                     mapData.remove(key);
-                    _DiwUtils.reply(cs, _ColorHelper.GREEN + "Deleted job: " + _ColorHelper.YELLOW + jobName);
+                    _DiwUtils.reply(player, _ColorHelper.GREEN + "Deleted job: " + _ColorHelper.YELLOW + jobName);
                     SaveData();
                 }
             }
         } else if (args.length >= 1 && args[0].equalsIgnoreCase("list")) {
             if (mapData.size() <= 0) {
-                _DiwUtils.reply(cs, _ColorHelper.AQUA + "There are no jobs setup.");
+                _DiwUtils.reply(player, _ColorHelper.AQUA + "There are no jobs setup.");
             } else {
-                _DiwUtils.reply(cs, _ColorHelper.LIGHT_PURPLE + RainbowUtils.TextLabel("Job Name", 10) + _ColorHelper.LIGHT_PURPLE + " " + RainbowUtils.TextLabel("Interval", 9) + _ColorHelper.LIGHT_PURPLE + " " + RainbowUtils.TextLabel("Time Left", 9) + _ColorHelper.LIGHT_PURPLE + " Command");
+                _DiwUtils.reply(player, _ColorHelper.LIGHT_PURPLE + RainbowUtils.TextLabel("Job Name", 10) + _ColorHelper.LIGHT_PURPLE + " " + RainbowUtils.TextLabel("Interval", 9) + _ColorHelper.LIGHT_PURPLE + " " + RainbowUtils.TextLabel("Time Left", 9) + _ColorHelper.LIGHT_PURPLE + " Command");
                 long msNow = System.currentTimeMillis();
-                List<String> keysSorted = new ArrayList<String>(mapData.keySet());
-                Collections.sort(keysSorted, new Comparator<String>() {
-                    public int compare(String o1, String o2) {
-                        _CronData data1 = (_CronData) _CmdCron.mapData.get(o1);
-                        _CronData data2 = (_CronData) _CmdCron.mapData.get(o2);
-                        return data1.msDelay > data2.msDelay ? 1 : (data1.msDelay < data2.msDelay ? -1 : 0);
-                    }
-                });
+                List<String> keysSorted = new ArrayList<>(mapData.keySet());
+                keysSorted.sort(Comparator.comparingLong(key -> _CmdCron.mapData.get(key).msDelay));
 
                 for (String key : keysSorted) {
-                    _CronData data = (_CronData) mapData.get(key);
+                    _CronData data = mapData.get(key);
                     String strTime = RainbowUtils.TimeDeltaString_JustMinutesSecs(data.msDelay);
                     long msLeft = data.msDelay - (msNow - data.msLastRun);
                     if (msLeft < 0L) {
@@ -209,66 +210,66 @@ public class _CmdCron extends CommandBase {
                     }
 
                     String strLeft = RainbowUtils.TimeDeltaString_JustMinutesSecs(msLeft);
-                    _DiwUtils.reply(cs, _ColorHelper.YELLOW + RainbowUtils.TextLabel(data.jobName, 10) + " " + _ColorHelper.WHITE + RainbowUtils.TextLabel(strTime, 9) + " " + _ColorHelper.GRAY + RainbowUtils.TextLabel(strLeft, 9) + " " + _ColorHelper.GOLD + data.cmdToRun);
+                    _DiwUtils.reply(player, _ColorHelper.YELLOW + RainbowUtils.TextLabel(data.jobName, 10) + " " + _ColorHelper.WHITE + RainbowUtils.TextLabel(strTime, 9) + " " + _ColorHelper.GRAY + RainbowUtils.TextLabel(strLeft, 9) + " " + _ColorHelper.GOLD + data.cmdToRun);
                 }
 
-                _DiwUtils.reply(cs, _ColorHelper.AQUA + mapData.size() + " jobs listed.");
+                _DiwUtils.reply(player, _ColorHelper.AQUA + mapData.size() + " jobs listed.");
             }
         } else if (args.length >= 1 && args[0].equalsIgnoreCase("setdelay")) {
             if (args.length < 3) {
-                _DiwUtils.reply(cs, _ColorHelper.RED + "Usage: /cron " + _ColorHelper.AQUA + "setdelay " + _ColorHelper.YELLOW + "MyLoop " + _ColorHelper.WHITE + "5m");
+                _DiwUtils.reply(player, _ColorHelper.RED + "Usage: /cron " + _ColorHelper.AQUA + "setdelay " + _ColorHelper.YELLOW + "MyLoop " + _ColorHelper.WHITE + "5m");
             } else {
                 String jobName = args[1].trim();
                 String strTime = RainbowUtils.ConcatArgs(args, 2).trim();
                 long ms = GetMSFromString(strTime);
                 if (ms <= 0L) {
-                    _DiwUtils.reply(cs, _ColorHelper.RED + "Usage: /cron " + _ColorHelper.AQUA + "setdelay " + _ColorHelper.YELLOW + "MyLoop " + _ColorHelper.WHITE + "5m");
+                    _DiwUtils.reply(player, _ColorHelper.RED + "Usage: /cron " + _ColorHelper.AQUA + "setdelay " + _ColorHelper.YELLOW + "MyLoop " + _ColorHelper.WHITE + "5m");
                 } else {
                     String key = jobName.toLowerCase();
-                    _CronData data = (_CronData) mapData.get(key);
+                    _CronData data = mapData.get(key);
                     if (data == null) {
-                        _DiwUtils.reply(cs, _ColorHelper.RED + "No job named: " + _ColorHelper.YELLOW + jobName);
+                        _DiwUtils.reply(player, _ColorHelper.RED + "No job named: " + _ColorHelper.YELLOW + jobName);
                     } else {
                         data.msDelay = ms;
                         mapData.put(key, data);
                         SaveData();
-                        _DiwUtils.reply(cs, _ColorHelper.GREEN + "Updated job " + _ColorHelper.AQUA + jobName + _ColorHelper.GREEN + " new interval: " + _ColorHelper.WHITE + RainbowUtils.TimeDeltaString_JustMinutesSecs(ms));
+                        _DiwUtils.reply(player, _ColorHelper.GREEN + "Updated job " + _ColorHelper.AQUA + jobName + _ColorHelper.GREEN + " new interval: " + _ColorHelper.WHITE + RainbowUtils.TimeDeltaString_JustMinutesSecs(ms));
                     }
                 }
             }
         } else if (args.length >= 1 && args[0].equalsIgnoreCase("setcmd")) {
             if (args.length < 3) {
-                _DiwUtils.reply(cs, _ColorHelper.RED + "Usage: /cron " + _ColorHelper.AQUA + "setcmd " + _ColorHelper.YELLOW + "MyLoop " + _ColorHelper.GOLD + "/give @a cookie");
+                _DiwUtils.reply(player, _ColorHelper.RED + "Usage: /cron " + _ColorHelper.AQUA + "setcmd " + _ColorHelper.YELLOW + "MyLoop " + _ColorHelper.GOLD + "/give @a cookie");
             } else {
                 String jobName = args[1].trim();
                 String strCmd = _DiwUtils.ConcatArgs(args, 2).trim();
                 if (strCmd.length() <= 0) {
-                    _DiwUtils.reply(cs, _ColorHelper.RED + "Usage: /cron " + _ColorHelper.AQUA + "setcmd " + _ColorHelper.YELLOW + "MyLoop " + _ColorHelper.GOLD + "/give @a cookie");
+                    _DiwUtils.reply(player, _ColorHelper.RED + "Usage: /cron " + _ColorHelper.AQUA + "setcmd " + _ColorHelper.YELLOW + "MyLoop " + _ColorHelper.GOLD + "/give @a cookie");
                 } else {
                     String key = jobName.toLowerCase();
-                    _CronData data = (_CronData) mapData.get(key);
+                    _CronData data = mapData.get(key);
                     if (data == null) {
-                        _DiwUtils.reply(cs, _ColorHelper.RED + "No job named: " + _ColorHelper.YELLOW + jobName);
+                        _DiwUtils.reply(player, _ColorHelper.RED + "No job named: " + _ColorHelper.YELLOW + jobName);
                     } else {
                         data.cmdToRun = strCmd;
                         mapData.put(key, data);
                         SaveData();
-                        _DiwUtils.reply(cs, _ColorHelper.GREEN + "Updated job " + _ColorHelper.AQUA + jobName + _ColorHelper.GREEN + " new command: " + _ColorHelper.GOLD + strCmd);
+                        _DiwUtils.reply(player, _ColorHelper.GREEN + "Updated job " + _ColorHelper.AQUA + jobName + _ColorHelper.GREEN + " new command: " + _ColorHelper.GOLD + strCmd);
                     }
                 }
             }
         } else if (args.length >= 2 && args[0].equalsIgnoreCase("advance")) {
             String strTime = RainbowUtils.ConcatArgs(args, 1).trim();
             long ms = GetMSFromString(strTime);
-            _DiwUtils.reply(cs, _ColorHelper.GREEN + "Advancing clock: " + _ColorHelper.WHITE + _DiwUtils.TimeDeltaString_NoDays(ms));
+            _DiwUtils.reply(player, _ColorHelper.GREEN + "Advancing clock: " + _ColorHelper.WHITE + _DiwUtils.TimeDeltaString_NoDays(ms));
 
             for (String key : mapData.keySet()) {
-                _CronData data = (_CronData) mapData.get(key);
+                _CronData data = mapData.get(key);
                 data.msLastRun -= ms;
             }
 
         } else {
-            this.ShowUsage(cs);
+            this.ShowUsage(player);
         }
     }
 }
