@@ -20,6 +20,11 @@ public enum MC_DamageType {
 	WITHER,
 	ANVIL,
 	FALLING_BLOCK,
+    HOT_FLOOR,
+    DRAGON_BREATH,
+    CRAMMING,
+    FLY_INTO_WALL,
+    FIREWORKS,
 
 	MOB,
 	PLAYER,
@@ -29,7 +34,5 @@ public enum MC_DamageType {
 	INDIRECT_MAGIC,
 	THORNS,
 	EXPLOSION,
-	EXPLOSION_PLAYER,
-	HOT_FLOOR,
-	DRAGON_BREATH
+	EXPLOSION_PLAYER
 }
