@@ -14,12 +14,15 @@ public enum MC_EnchantmentType
 	AQUA_AFFINITY,
 	THORNS,
 	DEPTH_STRIDER,
+	FROST_WALKER,
+	CURSE_OF_BINDING,
 	SHARPNESS,
 	SMITE,
 	BANE_OF_ARTHROPODS,
 	KNOCKBACK,
 	FIRE_ASPECT,
 	LOOTING,
+	SWEEPING,
 	EFFICIENCY,
 	SILK_TOUCH,
 	UNBREAKING,
@@ -30,7 +33,7 @@ public enum MC_EnchantmentType
 	INFINITY,
 	LUCK_OF_THE_SEA,
 	LURE,
-	UNKNOWN,
-	FROST_WALKER,
-	MENDING
+	MENDING,
+	CURSE_OF_VANISHING,
+	UNKNOWN
 }
