@@ -21,5 +21,8 @@ public enum MC_GameRuleType
 	REDUCED_DEBUG_INFO,
 	DO_ENTITY_DROPS,
 	SPECTATORS_GENERATE_CHUNKS,
-	DISABLE_ELYTRA_MOVEMENT_CHECK
+	DISABLE_ELYTRA_MOVEMENT_CHECK,
+	DO_WEATHER_CYCLE,
+	DO_LIMITED_CRAFTING,
+    ANNOUNCE_ADVANCEMENTS
 }
