@@ -1,44 +1,44 @@
 package org.projectrainbow.commands;
 
+import PluginReference.MC_Command;
 import PluginReference.MC_Player;
-import net.minecraft.command.CommandBase;
-import net.minecraft.command.CommandException;
-import net.minecraft.command.ICommandSender;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.server.MinecraftServer;
 import org.projectrainbow._ColorHelper;
 import org.projectrainbow.interfaces.IMixinEntityPlayerMP;
 
 import java.util.Collections;
 import java.util.List;
 
-public class _CmdBp extends CommandBase {
+public class _CmdBp implements MC_Command {
     @Override
     public String getCommandName() {
         return "bp";
     }
 
     @Override
-    public List<String> getCommandAliases() {
+    public List<String> getAliases() {
         return Collections.singletonList("backpack");
     }
 
     @Override
-    public boolean checkPermission(MinecraftServer minecraftServer, ICommandSender iCommandSender) {
-        return (!(iCommandSender instanceof MC_Player)) || ((MC_Player) iCommandSender).hasPermission("rainbow.bp");
+    public boolean hasPermissionToUse(MC_Player player) {
+        return player == null || player.hasPermission("rainbow.bp");
     }
 
     @Override
-    public String getCommandUsage(ICommandSender iCommandSender) {
+    public List<String> getTabCompletionList(MC_Player plr, String[] args) {
+        return Collections.emptyList();
+    }
+
+    @Override
+    public String getHelpLine(MC_Player player) {
         return _ColorHelper.AQUA + "/bp" + _ColorHelper.WHITE + " --- Backpack!";
     }
 
     @Override
-    public void execute(MinecraftServer minecraftServer, ICommandSender cs, String[] strings) throws CommandException {
-        EntityPlayer p = null;
-        if (cs instanceof EntityPlayer) {
-            p = (EntityPlayer) cs;
-            p.displayGUIChest(((IMixinEntityPlayerMP) p).getBackpack());
+    public void handleCommand(MC_Player player, String[] strings) {
+        if (player != null) {
+            ((EntityPlayer) player).displayGUIChest(((IMixinEntityPlayerMP) player).getBackpack());
         } else {
             System.out.println("--- Only for players!");
         }
