@@ -1,6 +1,11 @@
 package org.projectrainbow.mixins;
 
-import PluginReference.*;
+import PluginReference.MC_Entity;
+import PluginReference.MC_EventInfo;
+import PluginReference.MC_ItemStack;
+import PluginReference.MC_Location;
+import PluginReference.MC_Player;
+import PluginReference.MC_Sign;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.EntityPlayerMP;
@@ -11,8 +16,20 @@ import net.minecraft.nbt.NBTTagString;
 import net.minecraft.network.NetHandlerPlayServer;
 import net.minecraft.network.Packet;
 import net.minecraft.network.PacketBuffer;
-import net.minecraft.network.play.client.*;
-import net.minecraft.network.play.server.*;
+import net.minecraft.network.play.client.CPacketChatMessage;
+import net.minecraft.network.play.client.CPacketClickWindow;
+import net.minecraft.network.play.client.CPacketCustomPayload;
+import net.minecraft.network.play.client.CPacketPlayer;
+import net.minecraft.network.play.client.CPacketPlayerDigging;
+import net.minecraft.network.play.client.CPacketPlayerTryUseItem;
+import net.minecraft.network.play.client.CPacketPlayerTryUseItemOnBlock;
+import net.minecraft.network.play.client.CPacketUpdateSign;
+import net.minecraft.network.play.client.CPacketUseEntity;
+import net.minecraft.network.play.server.SPacketBlockChange;
+import net.minecraft.network.play.server.SPacketChat;
+import net.minecraft.network.play.server.SPacketConfirmTransaction;
+import net.minecraft.network.play.server.SPacketPlayerPosLook;
+import net.minecraft.network.play.server.SPacketSpawnPosition;
 import net.minecraft.server.management.PlayerList;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.tileentity.TileEntitySign;
@@ -22,12 +39,24 @@ import net.minecraft.util.IntHashMap;
 import net.minecraft.util.NonNullList;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Vec3d;
-import net.minecraft.util.text.*;
+import net.minecraft.util.text.ChatMessageType;
+import net.minecraft.util.text.ITextComponent;
+import net.minecraft.util.text.Style;
+import net.minecraft.util.text.TextComponentString;
+import net.minecraft.util.text.TextComponentTranslation;
+import net.minecraft.util.text.TextFormatting;
 import net.minecraft.util.text.event.ClickEvent;
 import net.minecraft.world.WorldServer;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.projectrainbow.*;
+import org.projectrainbow.GUIContainer;
+import org.projectrainbow.Hooks;
+import org.projectrainbow.PluginHelper;
+import org.projectrainbow._DiwUtils;
+import org.projectrainbow._DynReward;
+import org.projectrainbow._EmoteUtils;
+import org.projectrainbow._JOT_OnlineTimeUtils;
+import org.projectrainbow._JoeCommandStats;
 import org.projectrainbow.commands._CmdIgnore;
 import org.projectrainbow.interfaces.IMixinEntityPlayerMP;
 import org.projectrainbow.interfaces.IMixinOutboundPacketSoundEffect;
@@ -53,7 +82,7 @@ import java.util.regex.Pattern;
 public class MixinNetHandlerPlayServer {
 
     @Shadow
-    public EntityPlayerMP playerEntity;
+    public EntityPlayerMP player;
     @Shadow
     private int teleportId;
     @Shadow
@@ -67,57 +96,52 @@ public class MixinNetHandlerPlayServer {
     @Shadow
     private int networkTickCount;
 
-    @Redirect(method = "processPlayer", at = @At(value = "INVOKE", target = "warn", remap = false))
-    private void doLogWarning(Logger logger, String message, Object... args) {
+    @Redirect(method = "processPlayer", at = @At(value = "INVOKE", target = "org/apache/logging/log4j/Logger.warn(Ljava/lang/String;Ljava/lang/Object;)V ", remap = false))
+    private void doLogWarning(Logger logger, String message, Object arg) {
         if (!message.contains("is sending move packets too frequently") || !_DiwUtils.DoHideAnnoyingDefaultServerOutput) {
-            logger.warn(message, args);
+            logger.warn(message, arg);
         }
     }
 
     @Inject(method = "onDisconnect", at = @At("HEAD"))
     private void onDisconnect(ITextComponent var1, CallbackInfo callback) {
-        _JOT_OnlineTimeUtils.HandlePlayerLogout(playerEntity.getName(), playerEntity.getUniqueID());
-        Hooks.onPlayerLogout(playerEntity.getName(), playerEntity.getUniqueID());
-    }
-
-    @Redirect(method = "onDisconnect", at = @At(value = "INVOKE", target = "info", ordinal = 0, remap = false))
-    private void fixToString(Logger logger, String text, Object[] args, ITextComponent chatComponent) {
-        logger.info(this.playerEntity.getName() + " lost connection: " + chatComponent.getUnformattedText());
+        _JOT_OnlineTimeUtils.HandlePlayerLogout(player.getName(), player.getUniqueID());
+        Hooks.onPlayerLogout(player.getName(), player.getUniqueID());
     }
 
     @Inject(method = "processPlayerDigging", at = @At(value = "INVOKE", target = "net.minecraft.entity.player.EntityPlayerMP.markPlayerActive()V"), locals = LocalCapture.CAPTURE_FAILHARD, cancellable = true)
     private void onPacketPlayerDigging(CPacketPlayerDigging packet, CallbackInfo callback, WorldServer worldServer, BlockPos blockPos) {
         if (packet.getAction().equals(CPacketPlayerDigging.Action.START_DESTROY_BLOCK)) {
             MC_EventInfo ei = new MC_EventInfo();
-            Hooks.onAttemptBlockBreak((MC_Player) playerEntity, new MC_Location(blockPos.getX(), blockPos.getY(), blockPos.getZ(), playerEntity.dimension), ei);
+            Hooks.onAttemptBlockBreak((MC_Player) player, new MC_Location(blockPos.getX(), blockPos.getY(), blockPos.getZ(), player.dimension), ei);
             if (ei.isCancelled) {
-                playerEntity.connection.sendPacket(new SPacketBlockChange(worldServer, blockPos));
+                player.connection.sendPacket(new SPacketBlockChange(worldServer, blockPos));
                 callback.cancel();
             }
         }
     }
 
-    @Inject(method = "processRightClickBlock", at = @At(value = "INVOKE", target = "net.minecraft.entity.player.EntityPlayerMP.markPlayerActive()V"), locals = LocalCapture.CAPTURE_FAILHARD, cancellable = true)
+    @Inject(method = "processTryUseItemOnBlock", at = @At(value = "INVOKE", target = "net.minecraft.entity.player.EntityPlayerMP.markPlayerActive()V"), locals = LocalCapture.CAPTURE_FAILHARD, cancellable = true)
     private void processRightClickBlock(CPacketPlayerTryUseItemOnBlock packet, CallbackInfo callback, WorldServer worldServer,
                                         EnumHand hand, ItemStack itemStack, BlockPos blockPos,
                                         EnumFacing clickedFace) {
         MC_EventInfo ei = new MC_EventInfo();
-        Hooks.onAttemptPlaceOrInteract((MC_Player) playerEntity, new MC_Location(blockPos.getX(), blockPos.getY(), blockPos.getZ(), playerEntity.dimension), PluginHelper.directionMap.get(clickedFace), PluginHelper.handMap.get(hand), ei);
+        Hooks.onAttemptPlaceOrInteract((MC_Player) player, new MC_Location(blockPos.getX(), blockPos.getY(), blockPos.getZ(), player.dimension), PluginHelper.directionMap.get(clickedFace), PluginHelper.handMap.get(hand), ei);
         if (ei.isCancelled) {
             for (BlockPos pos : BlockPos.getAllInBox(blockPos.add(-1, -1, -1), blockPos.add(1, 1, 1))) {
                 if (pos.getY() < 0 || pos.getY() > 255) continue;
-                this.playerEntity.connection.sendPacket(new SPacketBlockChange(worldServer, pos));
+                this.player.connection.sendPacket(new SPacketBlockChange(worldServer, pos));
             }
             callback.cancel();
         } else {
-            _DynReward.HandleInteract((MC_Player) playerEntity, blockPos.getX(), blockPos.getY(), blockPos.getZ(), playerEntity.dimension);
+            _DynReward.HandleInteract((MC_Player) player, blockPos.getX(), blockPos.getY(), blockPos.getZ(), player.dimension);
         }
     }
 
-    @Inject(method = "processPlayerBlockPlacement", at = @At(value = "INVOKE", target = "net.minecraft.entity.player.EntityPlayerMP.markPlayerActive()V"), locals = LocalCapture.CAPTURE_FAILHARD, cancellable = true)
+    @Inject(method = "processTryUseItem", at = @At(value = "INVOKE", target = "net.minecraft.entity.player.EntityPlayerMP.markPlayerActive()V"), locals = LocalCapture.CAPTURE_FAILHARD, cancellable = true)
     public void onRightClickAir(CPacketPlayerTryUseItem packet, CallbackInfo callbackInfo, WorldServer world, EnumHand hand, ItemStack itemStack) {
         MC_EventInfo ei = new MC_EventInfo();
-        Hooks.onAttemptItemUse((MC_Player) playerEntity, (MC_ItemStack) (Object) itemStack, ei);
+        Hooks.onAttemptItemUse((MC_Player) player, (MC_ItemStack) (Object) itemStack, ei);
         if (ei.isCancelled) {
             callbackInfo.cancel();
         }
@@ -126,32 +150,32 @@ public class MixinNetHandlerPlayServer {
     @Inject(method = "processChatMessage", at = @At(value = "INVOKE", target = "net.minecraft.entity.player.EntityPlayerMP.markPlayerActive()V"), cancellable = true)
     public void onChat(CPacketChatMessage packet, CallbackInfo callbackInfo) {
         MC_EventInfo ei = new MC_EventInfo();
-        Hooks.onPlayerInput((MC_Player) playerEntity, packet.getMessage(), ei);
+        Hooks.onPlayerInput((MC_Player) player, packet.getMessage(), ei);
         if (ei.isCancelled) {
             callbackInfo.cancel();
         } else {
-            if (!((MC_Player) playerEntity).isOp()) {
-                String var11 = _JoeCommandStats.HandleNewCommand(playerEntity.getName(), packet.getMessage());
+            if (!((MC_Player) player).isOp()) {
+                String var11 = _JoeCommandStats.HandleNewCommand(player.getName(), packet.getMessage());
 
                 if (var11 != null) {
-                    ((MC_Player) playerEntity).kick(var11);
+                    ((MC_Player) player).kick(var11);
                     callbackInfo.cancel();
                     return;
                 }
 
                 if (_DiwUtils.HasBadLanguage(packet.getMessage())) {
-                    _DiwUtils.NotifyCensor(playerEntity.getName(), packet.getMessage());
+                    _DiwUtils.NotifyCensor(player.getName(), packet.getMessage());
                     long var14 = _DiwUtils.IncreaseEventCount("CENSOR." + packet.getMessage());
                     String swearMsg = String.format(
                             "Kicking for Language (Censor #%d): %s: %s",
-                            var14, packet.getMessage(), playerEntity.getName());
+                            var14, packet.getMessage(), player.getName());
 
                     System.out.println(
                             "--------------------------------------------");
                     System.out.println(swearMsg);
                     System.out.println(
                             "--------------------------------------------");
-                    ((MC_Player) playerEntity).kick(
+                    ((MC_Player) player).kick(
                             "Censor #" + var14
                                     + ": Watch language, Children play");
                     callbackInfo.cancel();
@@ -160,7 +184,7 @@ public class MixinNetHandlerPlayServer {
             }
 
             if (packet.getMessage().startsWith("/")) {
-                if (_EmoteUtils.HandleEmote((MC_Player) playerEntity, packet.getMessage().substring(1))) {
+                if (_EmoteUtils.HandleEmote((MC_Player) player, packet.getMessage().substring(1))) {
                     callbackInfo.cancel();
                     return;
                 }
@@ -168,19 +192,19 @@ public class MixinNetHandlerPlayServer {
         }
     }
 
-    @ModifyVariable(method = "processChatMessage", at = @At(value = "INVOKE", target = "startsWith", remap = false))
+    @ModifyVariable(method = "processChatMessage", at = @At(value = "INVOKE", target = "Ljava/lang/String;startsWith(Ljava/lang/String;)Z", remap = false))
     private String chatColor(String message) {
         if (message.startsWith("/") && !message.startsWith("/w") && !message.startsWith("/whisper") && !message.startsWith("/msg"))
             return message;
-        if (((MC_Player) playerEntity).hasPermission("rainbow.chatcolor")) {
-            return _DiwUtils.TranslateChatString(message, ((MC_Player) playerEntity).isOp());
+        if (((MC_Player) player).hasPermission("rainbow.chatcolor")) {
+            return _DiwUtils.TranslateChatString(message, ((MC_Player) player).isOp());
         }
         return message;
     }
 
     private static final Pattern patternLink = Pattern.compile("(?<=(\\W|^))(?:(https?)://)?([-\\w_\\.]{2,}\\.[a-z]{2,8})(/\\S*)?(?=(\\W|$))", Pattern.CASE_INSENSITIVE);
 
-    @Redirect(method = "processChatMessage", at = @At(value = "INVOKE", target = "net.minecraft.server.management.PlayerList.sendChatMsgImpl(Lnet/minecraft/util/text/ITextComponent;Z)V"))
+    @Redirect(method = "processChatMessage", at = @At(value = "INVOKE", target = "net.minecraft.server.management.PlayerList.sendMessage(Lnet/minecraft/util/text/ITextComponent;Z)V"))
     private void onChatSent(PlayerList playerList, ITextComponent message, boolean b) {
         TextComponentTranslation t = (TextComponentTranslation) message;
         ITextComponent player = (ITextComponent) t.getFormatArgs()[0];
@@ -208,11 +232,11 @@ public class MixinNetHandlerPlayServer {
         matcher.appendTail(buffer);
         result.appendSibling(new TextComponentString(buffer.toString()));
 
-        _DiwUtils.getMinecraftServer().addChatMessage(result);
+        _DiwUtils.getMinecraftServer().sendMessage(result);
 
-        SPacketChat packet = new SPacketChat(result, (byte) (b ? 1 : 0));
-        for (EntityPlayerMP receiver : playerList.getPlayerList()) {
-            if (!_CmdIgnore.IsIgnoring(receiver.getName(), playerEntity.getName())) {
+        SPacketChat packet = new SPacketChat(result, ChatMessageType.a((byte) (b ? 1 : 0)));
+        for (EntityPlayerMP receiver : playerList.getPlayers()) {
+            if (!_CmdIgnore.IsIgnoring(receiver.getName(), this.player.getName())) {
                 receiver.connection.sendPacket(packet);
             }
         }
@@ -222,50 +246,50 @@ public class MixinNetHandlerPlayServer {
     @Inject(method = "handleSlashCommand", at = @At("HEAD"))
     private void onCommand(String command, CallbackInfo callbackInfo) {
         if (!command.startsWith("/login")) {
-            Bootstrap.logger.info("" + ((MC_Player) playerEntity).getName() + ": " + command);
+            Bootstrap.logger.info("" + player.getName() + ": " + command);
         }
     }
 
     @Inject(method = "setPlayerLocation(DDDFFLjava/util/Set;)V", at = @At("HEAD"), cancellable = true)
     private void hookTeleport(double x, double y, double z, float yaw, float pitch, Set<SPacketPlayerPosLook.EnumFlags> flags, CallbackInfo callbackInfo) {
-        double dx = flags.contains(SPacketPlayerPosLook.EnumFlags.X) ? x : playerEntity.posX - x;
-        double dy = flags.contains(SPacketPlayerPosLook.EnumFlags.Y) ? y : playerEntity.posY - y;
-        double dz = flags.contains(SPacketPlayerPosLook.EnumFlags.Z) ? z : playerEntity.posZ - z;
-        float finalYaw = flags.contains(SPacketPlayerPosLook.EnumFlags.Y_ROT) ? yaw + playerEntity.rotationYaw : yaw;
-        float finalPitch = flags.contains(SPacketPlayerPosLook.EnumFlags.X_ROT) ? pitch + playerEntity.rotationPitch : pitch;
+        double dx = flags.contains(SPacketPlayerPosLook.EnumFlags.X) ? x : player.posX - x;
+        double dy = flags.contains(SPacketPlayerPosLook.EnumFlags.Y) ? y : player.posY - y;
+        double dz = flags.contains(SPacketPlayerPosLook.EnumFlags.Z) ? z : player.posZ - z;
+        float finalYaw = flags.contains(SPacketPlayerPosLook.EnumFlags.Y_ROT) ? yaw + player.rotationYaw : yaw;
+        float finalPitch = flags.contains(SPacketPlayerPosLook.EnumFlags.X_ROT) ? pitch + player.rotationPitch : pitch;
         // teleport
         MC_EventInfo ei = new MC_EventInfo();
-        Hooks.onAttemptPlayerTeleport((MC_Player) playerEntity, new MC_Location(playerEntity.posX + dx, playerEntity.posY + dy, playerEntity.posZ + dz, playerEntity.dimension, finalYaw, finalPitch), ei);
+        Hooks.onAttemptPlayerTeleport((MC_Player) player, new MC_Location(player.posX + dx, player.posY + dy, player.posZ + dz, player.dimension, finalYaw, finalPitch), ei);
         if (ei.isCancelled) {
             callbackInfo.cancel();
 
-            this.targetPos = new Vec3d(playerEntity.posX, playerEntity.posY, playerEntity.posZ);
+            this.targetPos = new Vec3d(player.posX, player.posY, player.posZ);
             if (++this.teleportId == 2147483647) {
                 this.teleportId = 0;
             }
             this.lastPositionUpdate = this.networkTickCount;
 
-            playerEntity.connection.sendPacket(new SPacketPlayerPosLook(playerEntity.posX, playerEntity.posY, playerEntity.posZ, playerEntity.rotationYaw, playerEntity.rotationPitch, Collections.<SPacketPlayerPosLook.EnumFlags>emptySet(), this.teleportId));
+            player.connection.sendPacket(new SPacketPlayerPosLook(player.posX, player.posY, player.posZ, player.rotationYaw, player.rotationPitch, Collections.emptySet(), this.teleportId));
         }
     }
 
     @Inject(method = "processPlayer", at = @At(value = "FIELD", target = "net.minecraft.network.NetHandlerPlayServer.serverController:Lnet/minecraft/server/MinecraftServer;", ordinal = 0), cancellable = true)
     private void onPlayerMove(CPacketPlayer packet, CallbackInfo callbackInfo) {
-        double x = packet.getX(this.playerEntity.posX);
-        double y = packet.getY(this.playerEntity.posY);
-        double z = packet.getZ(this.playerEntity.posZ);
-        float yaw = packet.getYaw(this.playerEntity.rotationYaw);
-        float pitch = packet.getPitch(this.playerEntity.rotationPitch);
-        double oldX = this.playerEntity.posX;
-        double oldY = this.playerEntity.posY;
-        double oldZ = this.playerEntity.posZ;
-        float oldYaw = this.playerEntity.rotationYaw;
-        float oldPitch = this.playerEntity.rotationPitch;
+        double x = packet.getX(this.player.posX);
+        double y = packet.getY(this.player.posY);
+        double z = packet.getZ(this.player.posZ);
+        float yaw = packet.getYaw(this.player.rotationYaw);
+        float pitch = packet.getPitch(this.player.rotationPitch);
+        double oldX = this.player.posX;
+        double oldY = this.player.posY;
+        double oldZ = this.player.posZ;
+        float oldYaw = this.player.rotationYaw;
+        float oldPitch = this.player.rotationPitch;
 
         if (Math.abs(x - oldX) > 1E-4 || Math.abs(y - oldY) > 1E-4 || Math.abs(z - oldZ) > 1E-4 || Math.abs(yaw - oldYaw) > 1E-4 || Math.abs(pitch - oldPitch) > 1E-4) {
 
             MC_EventInfo ei = new MC_EventInfo();
-            MC_Player player = (MC_Player) playerEntity;
+            MC_Player player = (MC_Player) this.player;
 
             MC_Location from = player.getLocation();
             MC_Location to = new MC_Location(x, y, z, from.dimension, yaw, pitch);
@@ -274,7 +298,7 @@ public class MixinNetHandlerPlayServer {
 
             if (ei.isCancelled) {
                 this.targetPos = new Vec3d(oldX, oldY, oldZ);
-                playerEntity.connection.sendPacket(new SPacketPlayerPosLook(oldX, oldY, oldZ, oldYaw, oldPitch, Collections.<SPacketPlayerPosLook.EnumFlags>emptySet(), ++this.teleportId));
+                this.player.connection.sendPacket(new SPacketPlayerPosLook(oldX, oldY, oldZ, oldYaw, oldPitch, Collections.emptySet(), ++this.teleportId));
                 callbackInfo.cancel();
             }
         }
@@ -285,9 +309,9 @@ public class MixinNetHandlerPlayServer {
         if (argPacket instanceof SPacketSpawnPosition) {
             BlockPos pos = ((IMixinOutboundPacketSpawnPosition) argPacket).getPos();
 
-            ((IMixinEntityPlayerMP) playerEntity).onCompassTargetUpdated(new MC_Location(
+            ((IMixinEntityPlayerMP) player).onCompassTargetUpdated(new MC_Location(
                     (double) pos.getX(), (double) pos.getY(),
-                    (double) pos.getZ(), this.playerEntity.dimension));
+                    (double) pos.getZ(), this.player.dimension));
         }
 
         if (argPacket instanceof IMixinOutboundPacketSoundEffect) {
@@ -295,10 +319,10 @@ public class MixinNetHandlerPlayServer {
             MC_EventInfo ei = new MC_EventInfo();
             MC_Location loc = new MC_Location(
                     (double) (packet.getX() / 8), (double) (packet.getY() / 8),
-                    (double) (packet.getZ() / 8), this.playerEntity.dimension);
+                    (double) (packet.getZ() / 8), this.player.dimension);
 
 
-            Hooks.onPacketSoundEffect((MC_Player) playerEntity, packet.getSoundName(), loc, ei);
+            Hooks.onPacketSoundEffect((MC_Player) player, packet.getSoundName(), loc, ei);
 
             if (ei.isCancelled) {
                 callbackInfo.cancel();
@@ -308,7 +332,7 @@ public class MixinNetHandlerPlayServer {
 
     @Redirect(method = "processUpdateSign", at = @At(value = "INVOKE", target = "net.minecraft.util.text.TextFormatting.getTextWithoutFormattingCodes(Ljava/lang/String;)Ljava/lang/String;"))
     private String stripColorCodes(String s) {
-        if (((MC_Player) playerEntity).hasPermission("rainbow.signcolor")) {
+        if (((MC_Player) player).hasPermission("rainbow.signcolor")) {
             return s;
         } else {
             return TextFormatting.getTextWithoutFormattingCodes(s);
@@ -318,7 +342,7 @@ public class MixinNetHandlerPlayServer {
     @Inject(method = "processUpdateSign", at = @At(value = "INVOKE", target = "net.minecraft.network.play.client.CPacketUpdateSign.getLines()[Ljava/lang/String;"), cancellable = true, locals = LocalCapture.CAPTURE_FAILHARD)
     private void onChangingSign(CPacketUpdateSign packet, CallbackInfo callbackInfo, WorldServer var2, BlockPos pos, IBlockState var4, TileEntity var5, TileEntitySign sign) {
         String[] newLines = packet.getLines();
-        MC_Player player = (MC_Player) playerEntity;
+        MC_Player player = (MC_Player) this.player;
 
         if (player.hasPermission("rainbow.signcolor")) {
             for (int i = 0; i < 4; ++i) {
@@ -328,9 +352,9 @@ public class MixinNetHandlerPlayServer {
 
         MC_EventInfo ei = new MC_EventInfo();
 
-        MC_Location location = new MC_Location(pos.getX(), pos.getY(), pos.getZ(), playerEntity.dimension);
+        MC_Location location = new MC_Location(pos.getX(), pos.getY(), pos.getZ(), this.player.dimension);
 
-        ArrayList<String> replaceLines = new ArrayList<String>();
+        ArrayList<String> replaceLines = new ArrayList<>();
 
         replaceLines.add(newLines[0]);
         replaceLines.add(newLines[1]);
@@ -352,7 +376,7 @@ public class MixinNetHandlerPlayServer {
                 replaceLines.get(0), replaceLines.get(1),
                 replaceLines.get(2), replaceLines.get(3));
         String var23 = String.format("------ [NEW SIGN by %s @ %s] --- %s",
-                this.playerEntity.getName(),
+                this.player.getName(),
                 player.getLocation().toString(), var22);
 
         LogManager.getLogger().debug(var23);
@@ -362,7 +386,7 @@ public class MixinNetHandlerPlayServer {
                 String line = newLines[i];
 
                 if (_DiwUtils.HasBadLanguage(line)) {
-                    _DiwUtils.NotifyCensor(this.playerEntity.getName(), "Sign: " + line);
+                    _DiwUtils.NotifyCensor(this.player.getName(), "Sign: " + line);
                     System.out.println("* Censoring Sign Line: " + line);
                     line = "Censored";
                     newLines[i] = line;
@@ -373,20 +397,20 @@ public class MixinNetHandlerPlayServer {
 
     @Inject(method = "processUpdateSign", at = @At(value = "INVOKE", target = "net.minecraft.tileentity.TileEntitySign.markDirty()V"), locals = LocalCapture.CAPTURE_FAILHARD)
     private void onChangedSign(CPacketUpdateSign packet, CallbackInfo callbackInfo, WorldServer var2, BlockPos pos, IBlockState var4, TileEntity var5, TileEntitySign sign) {
-        MC_Location location = new MC_Location(pos.getX(), pos.getY(), pos.getZ(), playerEntity.dimension);
+        MC_Location location = new MC_Location(pos.getX(), pos.getY(), pos.getZ(), player.dimension);
 
-        Hooks.onSignChanged((MC_Player) playerEntity, (MC_Sign) sign, location);
+        Hooks.onSignChanged((MC_Player) player, (MC_Sign) sign, location);
     }
 
     @Inject(method = "processCustomPayload", at = @At(value = "INVOKE", target = "net.minecraft.item.ItemStack.setTagInfo(Ljava/lang/String;Lnet/minecraft/nbt/NBTBase;)V", ordinal = 0), locals = LocalCapture.CAPTURE_FAILHARD, cancellable = true)
     private void onBookChange(CPacketCustomPayload packet, CallbackInfo callbackInfo, String var2, PacketBuffer var3, ItemStack newBook, ItemStack oldBook) {
         NBTTagList tagList = newBook.getTagCompound().getTagList("pages", 8);
-        ArrayList<String> pages = new ArrayList<String>();
+        ArrayList<String> pages = new ArrayList<>();
         for (int j = 0; j < tagList.tagCount(); j++) {
             pages.add(tagList.getStringTagAt(j));
         }
         MC_EventInfo ei = new MC_EventInfo();
-        Hooks.onAttemptBookChange((MC_Player) playerEntity, pages, ei);
+        Hooks.onAttemptBookChange((MC_Player) player, pages, ei);
         if (ei.isCancelled) {
             callbackInfo.cancel();
             return;
@@ -405,14 +429,14 @@ public class MixinNetHandlerPlayServer {
     @Inject(method = "processCustomPayload", at = @At(value = "INVOKE", target = "net.minecraft.item.ItemStack.setTagInfo(Ljava/lang/String;Lnet/minecraft/nbt/NBTBase;)V", ordinal = 3), locals = LocalCapture.CAPTURE_FAILHARD, cancellable = true)
     private void onBookSign(CPacketCustomPayload packet, CallbackInfo callbackInfo, String var2, PacketBuffer var3, ItemStack newBook, ItemStack oldBook) {
         NBTTagList tagList = newBook.getTagCompound().getTagList("pages", 8);
-        ArrayList<String> pages = new ArrayList<String>();
+        ArrayList<String> pages = new ArrayList<>();
         pages.add(oldBook.getTagCompound().getString("author"));
         pages.add(oldBook.getTagCompound().getString("title"));
         for (int j = 0; j < tagList.tagCount(); j++) {
             pages.add(tagList.getStringTagAt(j));
         }
         MC_EventInfo ei = new MC_EventInfo();
-        Hooks.onAttemptBookChange((MC_Player) playerEntity, pages, ei);
+        Hooks.onAttemptBookChange((MC_Player) player, pages, ei);
         if (ei.isCancelled) {
             callbackInfo.cancel();
             oldBook.getTagCompound().removeTag("author");
@@ -437,7 +461,7 @@ public class MixinNetHandlerPlayServer {
     private void onEntityInteract(CPacketUseEntity var1, CallbackInfo callbackInfo, WorldServer world, Entity target) {
         if (var1.getAction() == CPacketUseEntity.Action.INTERACT || var1.getAction() == CPacketUseEntity.Action.INTERACT_AT) {
             MC_EventInfo ei = new MC_EventInfo();
-            Hooks.onAttemptEntityInteract((MC_Player) playerEntity, (MC_Entity) target, ei);
+            Hooks.onAttemptEntityInteract((MC_Player) player, (MC_Entity) target, ei);
             if (ei.isCancelled) {
                 callbackInfo.cancel();
             }
@@ -446,31 +470,32 @@ public class MixinNetHandlerPlayServer {
 
     @Inject(method = "processClickWindow", at = @At(value = "INVOKE", target = "net.minecraft.inventory.Container.slotClick(IILnet/minecraft/inventory/ClickType;Lnet/minecraft/entity/player/EntityPlayer;)Lnet/minecraft/item/ItemStack;"), cancellable = true)
     private void onClickWindow(CPacketClickWindow packet, CallbackInfo ci) {
-        Container openContainer = playerEntity.openContainer;
+        Container openContainer = player.openContainer;
         if (openContainer instanceof GUIContainer) {
             pendingTransactions.addKey(openContainer.windowId, packet.getActionNumber());
-            playerEntity.connection.sendPacket(new SPacketConfirmTransaction(packet.getWindowId(), packet.getActionNumber(), false));
-            openContainer.setCanCraft(playerEntity, false);
-            openContainer.slotClick(packet.getSlotId(), packet.getUsedButton(), packet.getClickType(), playerEntity);
-            if (openContainer == playerEntity.openContainer) {
-                NonNullList<ItemStack> var8 = NonNullList.a();
+            player.connection.sendPacket(new SPacketConfirmTransaction(packet.getWindowId(), packet.getActionNumber(), false));
+            openContainer.setCanCraft(player, false);
+            openContainer.slotClick(packet.getSlotId(), packet.getUsedButton(), packet.getClickType(), player);
+            if (openContainer == player.openContainer) {
+                NonNullList<ItemStack> var8 = NonNullList.create();
 
-                for (int var4 = 0; var4 < this.playerEntity.openContainer.inventorySlots.size(); ++var4) {
-                    ItemStack var5 = this.playerEntity.openContainer.inventorySlots.get(var4).getStack();
+                for (int var4 = 0; var4 < this.player.openContainer.inventorySlots.size(); ++var4) {
+                    ItemStack var5 = this.player.openContainer.inventorySlots.get(var4).getStack();
                     ItemStack var6 = var5.isEmpty() ? ItemStack.EMPTY : var5;
+
                     var8.add(var6);
                 }
 
-                this.playerEntity.updateCraftingInventory(this.playerEntity.openContainer, var8);
+                this.player.updateCraftingInventory(this.player.openContainer, var8);
             }
             ci.cancel();
         }
     }
 
-    @Inject(method = "kickPlayerFromServer", at = @At("HEAD"), cancellable = true)
-    private void onKick(String reason, CallbackInfo callbackInfo) {
+    @Inject(method = "disconnect", at = @At("HEAD"), cancellable = true) // disconnect
+    private void onKick(ITextComponent reason, CallbackInfo callbackInfo) {
         MC_EventInfo ei = new MC_EventInfo();
-        Hooks.onPlayerKick((MC_Player) playerEntity, reason, ei);
+        Hooks.onPlayerKick((MC_Player) player, reason.getUnformattedText(), ei);
         if (ei.isCancelled) {
             callbackInfo.cancel();
         }
