@@ -2,7 +2,7 @@ package org.projectrainbow.plugins;
 
 import PluginReference.MC_Command;
 import PluginReference.MC_Player;
-import com.google.common.base.Objects;
+import com.google.common.base.MoreObjects;
 import net.minecraft.command.CommandBase;
 import net.minecraft.command.CommandException;
 import net.minecraft.command.ICommandSender;
@@ -21,7 +21,7 @@ public class PluginCommand extends CommandBase {
     }
 
     @Override
-    public String getCommandName() {
+    public String getName() {
         try {
             return delegate.getCommandName();
         } catch (Exception e) {
@@ -31,7 +31,7 @@ public class PluginCommand extends CommandBase {
     }
 
     @Override
-    public String getCommandUsage(ICommandSender sender) {
+    public String getUsage(ICommandSender sender) {
         try {
             return delegate.getHelpLine(sender instanceof MC_Player ? (MC_Player) sender : null);
         } catch (Exception e) {
@@ -50,12 +50,12 @@ public class PluginCommand extends CommandBase {
     }
 
     @Override
-    public List<String> getCommandAliases() {
+    public List<String> getAliases() {
         try {
-            return Objects.firstNonNull(delegate.getAliases(), Collections.<String>emptyList());
+            return MoreObjects.firstNonNull(delegate.getAliases(), Collections.<String>emptyList());
         } catch (Exception e) {
             e.printStackTrace();
-            return new ArrayList<String>();
+            return new ArrayList<>();
         }
     }
 
@@ -70,12 +70,12 @@ public class PluginCommand extends CommandBase {
     }
 
     @Override
-    public List<String> getTabCompletionOptions(MinecraftServer minecraftServer, ICommandSender sender, String[] args, BlockPos blockPos) {
+    public List<String> getTabCompletions(MinecraftServer minecraftServer, ICommandSender sender, String[] args, BlockPos blockPos) {
         try {
-            return Objects.firstNonNull(delegate.getTabCompletionList(sender instanceof MC_Player ? (MC_Player) sender : null, args), Collections.<String>emptyList());
+            return MoreObjects.firstNonNull(delegate.getTabCompletionList(sender instanceof MC_Player ? (MC_Player) sender : null, args), Collections.<String>emptyList());
         } catch (Exception e) {
             e.printStackTrace();
-            return new ArrayList<String>();
+            return new ArrayList<>();
         }
     }
 }
